backend/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-colorfield==0.6.3
Pillow==9.0.0
sqlparse==0.4.2
tzdata==2021.5